move sla.py to airflow/
